Добавляет вёрстку страницы 404
