+++ src/views/index.njk
@@ -1,4 +1,3 @@
-{% from "blocks/featured-article.njk" import featuredArticle %}
 {% from "blocks/header.njk" import header %}
 
 {% set isLogoContrastColor = hasCategory %}
@@ -21,27 +20,7 @@
       </div>
     </div>
 
-    <section class="articles-gallery">
-      <h2 class="visually-hidden">Избранные статьи</h2>
-
-      <ul class="featured-artices-list base-list">
-        {% for article in featuredArticles %}
-          {% set isHidden = loop.index > featuredArticlesPageSize %}
-          {% if isHidden %}
-            <template>
-              <li class="featured-artices-list__item">
-                {{ featuredArticle(article=article) }}
-              </li>
-            </template>
-          {% else %}
-            <li class="featured-artices-list__item">
-              {{ featuredArticle(article=article) }}
-            </li>
-          {% endif %}
-        {% endfor %}
-      </ul>
-      <button class="articles-gallery__more-button button" type="button">Показать ещё</button>
-    </section>
+    {% include "articles-gallery.njk" %}
 
     <section class="materials-collection">
       <h2 class="visually-hidden">Подборки материалов</h2>
